Added Notification System to rendercontroller Added check for "new controller detected" and "controller removed" with notification ScraperController now resize the image based on the current screen size (proportionally) Added a pointer to listview

+++ src/MVC/View/Image.cpp
@@ -69,7 +69,7 @@ void Image::resizeByHeight(float height)
 
 	if( tSize == Eigen::Vector2i::Zero()) throw "Invalid Operation, a texture is needed!";
 
-	float width = (int)(height * tSize.x() / tSize.y());
+	float width = height * tSize.x() / tSize.y();
 
 	setSize(width, height);
 }
